requirements.txt
pdfplumber>=0.9.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
anthropic>=0.7.0
google-generativeai>=0.3.0
scholarly>=1.7.0
bibtexparser>=1.4.0
jinja2>=3.1.0
pylatex>=1.4.0
pillow>=10.0.0
requests>=2.31.0
click>=8.1.0
google-genai>=0.1.0
